+++ packages/mobile/ios/NodeJsMobile/RNNodeJsMobile.m
@@ -6,6 +6,8 @@
 
 @implementation RNNodeJsMobile
 
+NSString* const EVENTS_CHANNEL = @"_EVENTS_";
+
 NSString* const BUILTIN_MODULES_RESOURCE_PATH = @"builtin_modules";
 NSString* const NODEJS_PROJECT_RESOURCE_PATH = @"nodejs-project";
 NSString* const NODEJS_DLOPEN_OVERRIDE_FILENAME = @"override-dlopen-paths-preload.js";
@@ -40,14 +42,6 @@ NSString* nodePath;
 }
 
 RCT_EXPORT_MODULE()
-
-RCT_EXPORT_METHOD(sendMessage:(NSString *)channelName:(NSString *)message)
-{
-  dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_BACKGROUND, 0), ^{
-    [[NodeRunner sharedInstance] sendMessageToNode:channelName:message];
-  });
-}
-
 
 -(void)callStartNodeProject:(NSString *)input
 {
@@ -102,6 +96,14 @@ RCT_EXPORT_METHOD(startNodeProject:(NSString *)command options:(NSDictionary *)o
     [nodejsThread setStackSize:2*1024*1024];
     [nodejsThread start];
   }
+}
+
+-(void)sendMessageToNode:(NSString *)event:(NSString *)message
+{
+  NSString * data = [NSString stringWithFormat:@"{ \"event\": \"%@\", \"payload\": \"%@\" }", event, message];
+  dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_BACKGROUND, 0), ^{
+    [[NodeRunner sharedInstance] sendMessageToNode:EVENTS_CHANNEL:data];
+  });
 }
 
 -(void) sendMessageBackToReact:(NSString*)channelName:(NSString*)message