+++ packages/mobile/src/screens/UsernameRegistration/UsernameRegistration.screen.tsx
@@ -3,7 +3,6 @@ import { useDispatch, useSelector } from 'react-redux'
 import { errors, identity } from '@quiet/state-manager'
 import { navigationActions } from '../../store/navigation/navigation.slice'
 import { ScreenNames } from '../../const/ScreenNames.enum'
-import { appImages } from '../../assets'
 import { UsernameRegistrationScreenProps } from './UsernameRegistration.types'
 import { UsernameRegistration } from '../../components/Registration/UsernameRegistration.component'
 import { ErrorCodes } from '@quiet/types'
@@ -30,23 +29,13 @@ export const UsernameRegistrationScreen: FC<UsernameRegistrationScreenProps> = (
     [dispatch]
   )
 
-  useEffect(() => {
-    if (usernameRegistered) {
-      navigation(ScreenNames.SuccessScreen, {
-        onPress: () => navigation(ScreenNames.ChannelListScreen),
-        icon: appImages.username_registered,
-        title: 'You created a username',
-        message: 'Your username will be registered shortly',
-      })
-    }
-  }, [currentIdentity?.userCertificate])
-
   const handleAction = (nickname: string) => {
     // Clear errors
     if (error) {
       dispatch(errors.actions.clearError(error))
     }
     dispatch(identity.actions.registerUsername(nickname))
+    navigation(ScreenNames.ConnectionProcessScreen)
   }
 
   return (