+++ packages/mobile/src/store/nativeServices/flushPersistor/flushPersistor.saga.ts
@@ -1,0 +1,7 @@
+import { call } from 'typed-redux-saga'
+import { persistor } from '../../store'
+
+export function* flushPersistorSaga(): Generator {
+  console.info('Flushing redux store')
+  yield* call(persistor.flush)
+}