+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.2-alpha.1](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@1.4.2-alpha.0...@quiet/mobile@1.4.2-alpha.1) (2023-07-17)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.4.2-alpha.0](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@1.4.1-alpha.2...@quiet/mobile@1.4.2-alpha.0) (2023-07-17)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.4.1-alpha.2](/compare/@quiet/mobile@1.4.1-alpha.1...@quiet/mobile@1.4.1-alpha.2) (2023-07-07)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.4.1-alpha.1](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.4.1-alpha.0...@quiet/mobile@1.4.1-alpha.1) (2023-07-07)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.4.1-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.4.0-alpha.6...@quiet/mobile@1.4.1-alpha.0) (2023-07-07)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
 # [1.4.0-alpha.6](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.4.0-alpha.5...@quiet/mobile@1.4.0-alpha.6) (2023-07-05)
 
 **Note:** Version bump only for package @quiet/mobile